requirements.txt
fastapi
uvicorn
google-generativeai
stripe
python-dotenv
faster-whisper
faiss-cpu
langchain
langchain-community
pydantic
sentence-transformers